chore(tar): Move errors into their own file so wrapper.cpp can use them.
